Moving all platform specific functionality into its own class.
All platform specific behaviour is isolated in its own class.
cli.py uses the `get_spotipy_class_by_platform` to decide which to use.
Also fixed the dbus interface previous on Linux. It was previously
using self.interface.Prev() but this fails out. It is actually
Previous(). I was using this website to determine this:
http://www.frandieguez.com/blog/2010/11/playing-with-d-bus-interface-of-spotify-for-linux/