Add prompt_toolkit v2 support in Pytify
prompt_toolkit v2 had a lot of breaking changes which caused Pytify to
crash when prompt_toolkit v2 was installed. This is an attempt to make
appropriate changes in Pytify to ensure Pytify does not break when
prompt_toolkit v2 is installed.

The following changes were made:
(i) PromptSession is used instead of prompt. This is the standard for
prompt_toolkit v2

(ii) AbortAction is removed in prompt_toolkit v2. Manual checks for
KeyboardInterrupt and EOF have to be done. on_abort parameter is removed
from the prompt function.

(iii) FileHistory does not have iter magic function anymore. a
load_history_strings() functions is provided which returns the list of
strings in the History file.

(iv) Token class is removed in v2. Styling of prompt is done by
pygments. Style class is used from prompt_toolkit.styles to define the
styling of the prompt.

(v) complete_in_thread is added to PromptSession to enable asynchronous
completion of text.

(vi) WordCompleter accepts a list of words rather than a set of words. A
list of words is passed as the first param to WordCompleter rather thana
set.

(vii) get_prompt_tokens and get_bottom_toolbar_tokens were renamed to
get_prompt and get_bottom. This is because tokens are not
supported anymore by prompt_toolkit, the naming should be appropriate to
reflect this.

This code was tested by running all the possible commands.